requirements.txt
python-dotenv==1.1.0
slack_sdk==3.35.0
